Add first triangle
Glew had to be added, because from OpenGL 1.1 the pointers to GPU funcs
and definitions have to be set - not included with gl.h (which is
seemingly part of wxGLCanvas). In pyramid example it is done manually,
but can be done in bulk using Glew.
OpenGL is separated from wxWidgets code (OGL is only in graphics files).
This is the first attempt at using OpenGL - there is only the triangle
example as a placeholder almost without any structures - just to prove
everything works. The viewport interferes with the interface.