Add render loop
Previously the OpenGL would render a frame only when the application
started or the window was resized. Now the image is refreshed regularly.
To demonstrate this, the tetrahedron changes color over time - using
a uniform to input the color to the fragment shader.
Makefile now compiles in parallel and includes extra warning messages.