Rework debugging
Previously were errors handled with glGetError being manually called
after each OpenGL call, using enum to determine the cause of the error.
Now it is done automatically using a callback function and KHR_debug
extension.
The callback function is sourced from Github, because it is basically
just print function with switch statements so it is readable.
Added checks for this and ARB_direct_state_access
extension, which was already used.
Fixed the last memory leak