Add object class
Currently the object cannot have texture assigned to it, because it
should be handled in a texture management menu, which doesn't yet exist.
Every object has its own vertex array and is rendered in a separate
OpenGL draw call, which is truly not very much optimized approach,
but it is the easiest and cleanest.