Rework render loop
Previously was render loop with a wxTimer set on 0 duration repeating
infinitely. This solution is not ideal on Windows, because wxTimer
relies on low precision timer, which in Windows measures in 16ms steps.
So it was not possible to reach higher frame rate than ~60Hz.
Now the rendering is done with a event, which invokes itself. This
approach basically breaks the original main loop, so the better way to
do this is probably modifying the main loop itself.