Start expanding the UI and little fixes
KHR_debug message callback now always outputs to std::cout.
Fixed some memory leaks.
I was planning to rework the current render loop - modify the event loop
itself instead of making something like it using just events and the
current event loop. But the wxEventLoop is very poorly documented, so I
was not able to. Overriding the Run() seemed it did not do anything,
even though it is a virtual method just for this purpose.
I tried to make the UI using multiple frames and wxLayoutAlgorithm,
but doing it with sizers seems much easier with almost the same effect.
(Also implementation of this class is kind of documented in comparison
with the layout algorithm.)
The only disadvantage is that user cannot move the UI elements.