Remove TextureManager class and change textures
TextureManager only held the textures in a vector - now it is done by
GraphicsManager with almost no difference.
Previously the textures were shown using the ARB_bindless_texture
extension, but this extension appears either to be bugged or has to be
used differently with Intel GPU drivers (the textures did not work with
Intel Iris XE). The bindless textures could be used to render all
objects in a single draw call, but that is not the way this app works
(it is a better way, but more complicated).
So it seemingly makes very small difference.