Fix object loading and improve makefile
Cleaning using makefile now works even if some of the files are still
present and some deleted.
Previously face data could be parsed incorrectly - normals were
interpreted as texture coordinates.
When the object file is corrupted, the loader now is able to to
recognize and report to user if there is incorrect number of parameters
on a line or the index is of the vertex, texture or normal is out of
range.